requirements.txt
python-dotenv==1.0.1
streamlit==1.32.2
PyMuPDF==1.25.2
openai==1.16.1